+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support user file uploads to S3 [#114](https://github.com/rokwire/content-building-block/issues/114)
 - Add POST /content-items API
 - Separate S3 signed URL expiration environment variables
+- Support S3 multipart uploads
 ### Changed
 - Generate file IDs for S3 file uploads
 - Reintroduce file extensions for S3